move version into new package.
