+++ Makefile
@@ -18,7 +18,7 @@ endif
 # See pkg/cmd/version.go for details
 SOURCE_GIT_COMMIT ?= $(shell git rev-parse --short HEAD)
 BUILD_VERSION ?= $(shell git describe --tags --abbrev=40 --dirty)
-VERSION_URI = "github.com/nitrictech/cli/pkg/utils"
+VERSION_URI = "github.com/nitrictech/cli/pkg/version"
 export LDFLAGS="-X $(VERSION_URI).Version=${BUILD_VERSION} \
                 -X $(VERSION_URI).Commit=${SOURCE_GIT_COMMIT} \
                 -X $(VERSION_URI).BuildTime=$(shell date +%Y-%m-%dT%H:%M:%S%z)"
